# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface FoodSearchItem {
  id: string;
  name: string;
  serving: string;
  calories: number;
  protein: number;
  carbs: number;
  fat: number;
}

export interface FoodSearchResponse {
  query: string;
  items: FoodSearchItem[];
}

export interface ErrorResponse {
  error: string;
}

export interface RunForgeDiscountClaimRequest {
  /**
     * @minLength 1
     * @maxLength 200
     */
  clientId: string;
}

export interface RunForgeDiscountResponse {
  available: boolean;
  /** @nullable */
  code: string | null;
  /** @minimum 0 */
  remaining: number;
  /** @minimum 1 */
  limit: number;
}

export type SearchFoodParams = {
/**
 * Food or drink search text
 * @minLength 2
 */
q: string;
/**
 * Preferred product name language
 */
language?: SearchFoodLanguage;
/**
 * Maximum number of results
 * @minimum 1
 * @maximum 20
 */
limit?: number;
};

export type SearchFoodLanguage = typeof SearchFoodLanguage[keyof typeof SearchFoodLanguage];


export const SearchFoodLanguage = {
  tr: 'tr',
  en: 'en',
  de: 'de',
  fr: 'fr',
  es: 'es',
} as const;