# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: food
    description: Food search operations
  - name: runforge
    description: RunForge discount operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /food/search:
    get:
      operationId: searchFood
      tags: [food]
      summary: Search food nutrition
      description: Searches a broad food database and returns normalized nutrition values.
      parameters:
        - name: q
          in: query
          required: true
          description: Food or drink search text
          schema:
            type: string
            minLength: 2
        - name: language
          in: query
          required: false
          description: Preferred product name language
          schema:
            type: string
            enum: [tr, en, de, fr, es]
            default: en
        - name: limit
          in: query
          required: false
          description: Maximum number of results
          schema:
            type: integer
            minimum: 1
            maximum: 20
            default: 12
      responses:
        "200":
          description: Normalized food search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FoodSearchResponse"
        "400":
          description: Invalid search query
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Food data provider unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /runforge/discount:
    post:
      operationId: claimRunForgeDiscount
      tags: [runforge]
      summary: Claim the shared RunForge discount code
      description: Assigns the shared RunForge discount code to one unique client within the campaign limit.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RunForgeDiscountClaimRequest"
      responses:
        "200":
          description: Discount code claim status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunForgeDiscountResponse"
        "400":
          description: Invalid client ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Discount allocation unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    FoodSearchItem:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        serving:
          type: string
        calories:
          type: number
        protein:
          type: number
        carbs:
          type: number
        fat:
          type: number
      required:
        - id
        - name
        - serving
        - calories
        - protein
        - carbs
        - fat
    FoodSearchResponse:
      type: object
      properties:
        query:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/FoodSearchItem"
      required:
        - query
        - items
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    RunForgeDiscountClaimRequest:
      type: object
      properties:
        clientId:
          type: string
          minLength: 1
          maxLength: 200
      required:
        - clientId
    RunForgeDiscountResponse:
      type: object
      properties:
        available:
          type: boolean
        code:
          type:
            - string
            - "null"
        remaining:
          type: number
          minimum: 0
        limit:
          type: number
          minimum: 1
      required:
        - available
        - code
        - remaining
        - limit