# lib/api-zod/src/generated/types/runForgeDiscountResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RunForgeDiscountResponse {
  available: boolean;
  /** @nullable */
  code: string | null;
  /** @minimum 0 */
  remaining: number;
  /** @minimum 1 */
  limit: number;
}